Setup Converters Package properly for Sample App Usage
Before it meant that we couldn't work on the Converters component and see results in Sample App as it was pulling in the pre-built NuGet package
With this change, we can define the property in the root props for the Windows repo (like we do with Extensions) so that it'll use the source project when running from there. Like Extensions does for the Tests.