Another fix for the Converters ProjectReference logic
We need to add the Converters when running in Single Component mode,
as it's only included in the all-up sample app locally (when source defined).

We also need to exclude if we're building the Converters project itself.

However, if a component itself relies on the converters and we try and to build
on Linux we'll hit the original issue. Workaround for now there is to build
all-up sample app instead...

Tested locally that the Converters and another component both could be built now.